[refactor] added interfaces directory module
