+++ src/echo_vision/interfaces/capturer_interface.py
